Reimplemented AssemblyInfo to allow editing and saving
Original file content is saved in the AssemblyInfo as a list of lines.

Each parsed entity (title, versions) is saved with the line of the original
content it is found in.

RegexExtractor.replace is used to edit lines corresponding to parsed entities.

FileExtractor.save is used to save the content back to the original file on
disk.